move view to separate file
