+++ jvm/workbookapp/src/main/kotlin/org/wycliffeassociates/otter/jvm/workbookapp/ui/screens/ChunkingTranslationPage.kt
@@ -23,6 +23,7 @@ import org.wycliffeassociates.otter.jvm.workbookapp.ui.screens.translation.Chunk
 import org.wycliffeassociates.otter.jvm.workbookapp.ui.screens.translation.Consume
 import org.wycliffeassociates.otter.jvm.workbookapp.ui.screens.translation.PeerEdit
 import org.wycliffeassociates.otter.jvm.workbookapp.ui.screens.translation.ChapterReview
+import org.wycliffeassociates.otter.jvm.workbookapp.ui.screens.translation.SourceAudioMissing
 import org.wycliffeassociates.otter.jvm.workbookapp.ui.screens.translation.translationHeader
 import org.wycliffeassociates.otter.jvm.workbookapp.ui.viewmodel.TranslationViewModel2
 import org.wycliffeassociates.otter.jvm.workbookapp.ui.viewmodel.WorkbookDataStore
@@ -34,21 +35,18 @@ class ChunkingTranslationPage : View() {
     val viewModel: TranslationViewModel2 by inject()
     val workbookDataStore: WorkbookDataStore by inject()
 
-    private val sourceAudioMissingView = createSourceAudioMissingView()
-
-    private val mainFragmentProperty = viewModel.selectedStepProperty.objectBinding {
-        it?.let { step ->
-            val fragment = when(step) {
-                ChunkingStep.CONSUME_AND_VERBALIZE -> find<Consume>()
-                ChunkingStep.CHUNKING -> find<Chunking>()
-                ChunkingStep.BLIND_DRAFT -> find<BlindDraft>()
-                ChunkingStep.PEER_EDIT,
-                ChunkingStep.KEYWORD_CHECK,
-                ChunkingStep.VERSE_CHECK -> find<PeerEdit>()
-                ChunkingStep.FINAL_REVIEW -> find<ChapterReview>()
-            }
-            fragment.root
-        } ?: sourceAudioMissingView
+    private val mainFragmentProperty = viewModel.selectedStepProperty.objectBinding { step ->
+        val fragment = when(step) {
+            ChunkingStep.CONSUME_AND_VERBALIZE -> find<Consume>()
+            ChunkingStep.CHUNKING -> find<Chunking>()
+            ChunkingStep.BLIND_DRAFT -> find<BlindDraft>()
+            ChunkingStep.PEER_EDIT,
+            ChunkingStep.KEYWORD_CHECK,
+            ChunkingStep.VERSE_CHECK -> find<PeerEdit>()
+            ChunkingStep.FINAL_REVIEW -> find<ChapterReview>()
+            null -> find<SourceAudioMissing>()
+        }
+        fragment.root
     }
 
     private lateinit var sourceTextDrawer: SourceTextDrawer
@@ -133,125 +131,5 @@ class ChunkingTranslationPage : View() {
     override fun onUndock() {
         super.onUndock()
         viewModel.undockPage()
-    }
-
-    private fun createSourceAudioMissingView(): Node {
-        return VBox().apply {
-            addClass("audio-missing-view")
-            vgrow = Priority.ALWAYS
-            visibleWhen { viewModel.showAudioMissingViewProperty }
-            managedWhen(visibleProperty())
-
-            label {
-                addClass("audio-missing__title")
-                textProperty().bind(
-                    workbookDataStore.activeChapterTitleBinding().stringBinding {
-                        MessageFormat.format(
-                            messages["source_audio_missing_for"],
-                            it
-                        )
-                    }
-                )
-            }
-            label(messages["source_audio_missing_description"]) {
-                addClass("normal-text", "audio-missing__description")
-            }
-
-            vbox {
-                addClass("audio-missing__drag-drop-area")
-
-                label {
-                    graphic = FontIcon(MaterialDesign.MDI_FOLDER_OUTLINE).addClass("big-icon")
-                }
-
-                textflow {
-                    addClass("audio-missing__text-flow")
-                    textAlignment = TextAlignment.CENTER
-
-                    val textMessage = messages["drag_drop_or_browse_import__template"]
-                    val prefixText = textMessage.substringBefore('{')
-                    val suffixText = textMessage.substringAfter('}')
-
-                    text(prefixText) {
-                        addClass("normal-text", "centered")
-                    }
-                    hyperlink(messages["choose_file"]).apply {
-                        addClass("wa-text--hyperlink", "audio-missing__link-text")
-                        tooltip(text)
-                        action {
-                            chooseFile(
-                                FX.messages["importResourceFromZip"],
-                                arrayOf(
-                                    FileChooser.ExtensionFilter(
-                                        messages["oratureFileTypes"],
-                                        *OratureFileFormat.extensionList.map { "*.$it" }.toTypedArray()
-                                    )
-                                ),
-                                mode = FileChooserMode.Single,
-                                owner = currentWindow
-                            )//.firstOrNull()?.let { importFile(it) }
-                        }
-                    }
-                    text(suffixText) {
-                        addClass("normal-text", "centered")
-                    }
-                }
-            }
-
-            stackpane {
-                addClass("audio-missing__separator-area")
-                separator { fitToParentWidth() }
-            }
-
-            label(messages["need_source_audio"]) {
-                addClass("h3")
-            }
-
-            textflow {
-                addClass("audio-missing__text-flow")
-
-                val textMessage = messages["source_audio_download_description__template"]
-                val prefixText = textMessage.substringBefore('{')
-                val suffixText = textMessage.substringAfter('}')
-                text(prefixText) {
-                    addClass("normal-text")
-                }
-                hyperlink("audio.bibleineverylanguage.org").apply {
-                    addClass("wa-text--hyperlink", "audio-missing__link-text")
-                    tooltip("audio.bibleineverylanguage.org/gl")
-                    action {
-                        hostServices.showDocument("https://audio.bibleineverylanguage.org/gl")
-                    }
-                }
-                text(suffixText) {
-                    addClass("normal-text")
-                }
-            }
-
-            hbox {
-                addClass("audio-missing__actions")
-
-                button(messages["check_online"]) {
-                    addClass("btn", "btn--primary")
-                    tooltip(text)
-                    graphic = FontIcon(MaterialDesign.MDI_EXPORT)
-                    prefWidthProperty().bind(this@hbox.widthProperty().divide(2))
-                    action {
-                        hostServices.showDocument("https://audio.bibleineverylanguage.org/gl")
-                    }
-                }
-
-                button {
-                    addClass("btn", "btn--secondary")
-                    text = MessageFormat.format(
-                        messages["begin_narrating_book"],
-                        viewModel.bookTitleProperty.value
-                    )
-                    tooltip(text)
-                    graphic = FontIcon(MaterialDesign.MDI_ARROW_RIGHT)
-                    prefWidthProperty().bind(this@hbox.widthProperty().divide(2))
-                }
-            }
-        }
     }
 }
