+++ jvm/workbookapp/src/main/resources/Messages_en.properties
@@ -452,7 +452,7 @@ chunkingDescription = Break the passage of Scripture into smaller units or parts
 # Example: Source Audio Missing for Chapter 2
 source_audio_missing_for = Source Audio Missing for {0}
 source_audio_missing_description = Orature uses source audio to complete several steps of the MAST process. Please import audio to continue.
-drag_drop_or_browse_import__template = Drag and Drop or {0} to upload
+drag_drop_or_browse_import__template = Drag and Drop or {0} to import
 need_source_audio = Need Source Audio?
 source_audio_download_description__template = Source audio for this project may be found online at {0}. Source audio may not always be available online but can be created by starting a new narration project.
 check_online = Check Online
